requirements.txt
nltk~=3.8.1
PyPDF2~=3.0.1
spacy~=3.7.5
pandas~=2.2.2
yake~=0.4.8
scikit-learn~=1.5.1
fuzzywuzzy~=0.18.0
PyMuPDF~=1.24.9
python-dotenv~=1.0.1
